Maryia/extracted StepNavigation into a separate file
